SampleKotlin프로젝트에서 작업을 했어야 했는데 SunFlower-MySelf-Kotlin 프로젝트에서 작업을 했다. 작업한 내용을 겨우 복붙해서 이 프로젝트에 담은후 기능이 정상 동작되는 것 까지 확인했다.
