+++ app/src/main/AndroidManifest.xml
@@ -10,10 +10,10 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.SampleKotlin"
+        android:theme="@style/Theme.SunFlowerMyselfKotlin"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".adpater.view.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
