--- app/src/main/java/com/example/samplekotlin/sunflower_myself_kotlin/vo/Plant.kt
@@ -1,6 +1,0 @@
-package com.example.sunflower_myself_kotlin.vo
-
-data class Plant(val id : Long = 0, val name : String, val waterPeriod : Int, val imageResource : Int){
-
-}
-
