+++ app/src/main/java/com/example/samplekotlin/vo/Plant.kt
@@ -1,0 +1,6 @@
+package com.example.samplekotlin.vo
+
+data class Plant(val id : Long = 0, val name : String, val waterPeriod : Int, val imageResource : Int){
+
+}
+
