+++ app/src/main/res/drawable/add_btn.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/white"></solid>
+
+    <corners
+        android:bottomLeftRadius="15dp"
+        android:topRightRadius="15dp"
+        />
+
+
+    <size
+        android:width="110dp"
+        android:height="40dp" />
+
+</shape>
