+++ app/src/main/res/drawable/ic_my_garden_active.xml
@@ -1,0 +1,23 @@
+<!--
+  ~ Copyright 2019 Google LLC
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~     https://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<vector android:height="28dp" android:viewportHeight="21"
+    android:viewportWidth="18" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path
+        android:pathData="M9,21C13.97,21 18,16.97 18,12C13.03,12 9,16.03 9,21ZM2.6,9.25C2.6,10.63 3.72,11.75 5.1,11.75C5.63,11.75 6.11,11.59 6.52,11.31L6.5,11.5C6.5,12.88 7.62,14 9,14C10.38,14 11.5,12.88 11.5,11.5L11.48,11.31C11.88,11.59 12.37,11.75 12.9,11.75C14.28,11.75 15.4,10.63 15.4,9.25C15.4,8.25 14.81,7.4 13.97,7C14.81,6.6 15.4,5.75 15.4,4.75C15.4,3.37 14.28,2.25 12.9,2.25C12.37,2.25 11.89,2.41 11.48,2.69L11.5,2.5C11.5,1.12 10.38,0 9,0C7.62,0 6.5,1.12 6.5,2.5L6.52,2.69C6.12,2.41 5.63,2.25 5.1,2.25C3.72,2.25 2.6,3.37 2.6,4.75C2.6,5.75 3.19,6.6 4.03,7C3.19,7.4 2.6,8.25 2.6,9.25ZM9,4.5C10.38,4.5 11.5,5.62 11.5,7C11.5,8.38 10.38,9.5 9,9.5C7.62,9.5 6.5,8.38 6.5,7C6.5,5.62 7.62,4.5 9,4.5ZM0,12C0,16.97 4.03,21 9,21C9,16.03 4.97,12 0,12Z"
+        android:fillColor="?attr/colorOnPrimary"
+        android:fillType="nonZero" />
+</vector>
