+++ app/src/main/res/drawable/rounded_corner_image.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/white"></solid>
+    <corners android:topRightRadius="15dp"/>
+
+    <size
+        android:width="110dp"
+        android:height="40dp" />
+</shape>
