+++ app/src/main/res/menu/top_nav_menu.xml
@@ -1,0 +1,10 @@
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <item
+        android:id="@+id/settings"
+        android:title="Sunflowerzzz"
+        android:icon="@drawable/ic_filter_list_24dp"
+        app:showAsAction="always"
+        />
+
+</menu>
