+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,6 @@
 <resources>
-    <string name="app_name">SampleKotlin</string>
+    <string name="app_name">SunFlower-Myself-Kotlin</string>
+    <string name="Sunflower">Sunflower</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
