+++ app/src/main/res/values/style.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<resources>
+    <style name="font_set_a">
+        <item name="android:textColor">@color/black</item>
+        <item name="android:textSize">36px</item>
+    </style>
+    <style name="font_set_b">
+        <item name="android:textColor">@color/black</item>
+        <item name="android:textSize">24px</item>
+    </style>
+</resources>
